@@ -6,30 +6,30 @@
  * @license    http://www.opensource.org/licenses/mit-license.html
  * @link       http://antecedent.github.com/patchwork
  */
-namespace Patchwork\Preprocessor\Callbacks\Interceptor;
+namespace Patchwork\CodeManipulation\Actions\CodeManipulation;
 
-use Patchwork\Preprocessor\Callbacks\Generic;
-use Patchwork\Interceptor;
+use Patchwork\CodeManipulation\Actions\Generic;
+use Patchwork\CallRerouting;
 use Patchwork\Utils;
 
 const CALL_INTERCEPTION_CODE = '
-    $__pwClosureName = __NAMESPACE__ ? __NAMESPACE__ . "\\{closure}" : "{closure}";
+    $__pwClosureName = __NAMESPACE__ ? __NAMESPACE__ . "\\\\{closure}" : "\\\\{closure}";
     $__pwClass = (__CLASS__ && __FUNCTION__ !== $__pwClosureName) ? __CLASS__ : null;
-    if (!empty(\Patchwork\Interceptor\State::$patches[$__pwClass][__FUNCTION__])) {
+    if (!empty(\Patchwork\CallRerouting\State::$routes[$__pwClass][__FUNCTION__])) {
         $__pwCalledClass = $__pwClass ? \get_called_class() : null;
         $__pwFrame = \count(\debug_backtrace(false));
-        if (\Patchwork\Interceptor\intercept($__pwClass, $__pwCalledClass, __FUNCTION__, $__pwFrame, $__pwResult)) {
+        if (\Patchwork\CallRerouting\dispatch($__pwClass, $__pwCalledClass, __FUNCTION__, $__pwFrame, $__pwResult)) {
             return $__pwResult;
         }
     }
     unset($__pwClass, $__pwCalledClass, $__pwResult, $__pwClosureName, $__pwFrame);
 ';
 
-const QUEUE_DEPLOYMENT_CODE = '\Patchwork\Interceptor\deployQueue()';
+const QUEUE_DEPLOYMENT_CODE = '\Patchwork\CallRerouting\deployQueue()';
 
 function markPreprocessedFiles()
 {
-    return Generic\markPreprocessedFiles(Interceptor\State::$preprocessedFiles);
+    return Generic\markPreprocessedFiles(CallRerouting\State::$preprocessedFiles);
 }
 
 function injectCallInterceptionCode()
