@@ -8,6 +8,13 @@
  */
 namespace Patchwork\Utils;
 
+const ALIASING_CODE = '
+    namespace %s;
+    function %s() {
+        return call_user_func_array("%s", func_get_args());
+    }
+';
+
 function clearOpcodeCaches()
 {
     if (ini_get('wincache.ocenabled')) {
@@ -16,11 +23,6 @@ function clearOpcodeCaches()
     if (ini_get('apc.enabled')) {
         apc_clear_cache();
     }
-}
-
-function traitsSupported()
-{
-    return version_compare(PHP_VERSION, "5.4", ">=");
 }
 
 function generatorsSupported()
@@ -35,7 +37,7 @@ function runningOnHHVM()
 
 function condense($string)
 {
-    return preg_replace("/\s*/", "", $string);
+    return preg_replace('/\s*/', '', $string);
 }
 
 function findFirstGreaterThan(array $array, $value)
@@ -46,7 +48,7 @@ function findFirstGreaterThan(array $array, $value)
         return $high + 1;
     }
     while ($low < $high) {
-        $mid = (int) (($low + $high) / 2);
+        $mid = (int)(($low + $high) / 2);
         if ($array[$mid] <= $value) {
             $low = $mid + 1;
         } else {
@@ -56,10 +58,10 @@ function findFirstGreaterThan(array $array, $value)
     return $low;
 }
 
-function interpretCallback($callback)
+function interpretCallable($callback)
 {
     if (is_object($callback)) {
-        return interpretCallback(array($callback, "__invoke"));
+        return interpretCallable([$callback, "__invoke"]);
     }
     if (is_array($callback)) {
         list($class, $method) = $callback;
@@ -69,19 +71,19 @@ function interpretCallback($callback)
             $class = get_class($class);
         }
         $class = ltrim($class, "\\");
-        return array($class, $method, $instance);
+        return [$class, $method, $instance];
     }
     $callback = ltrim($callback, "\\");
     if (strpos($callback, "::")) {
         list($class, $method) = explode("::", $callback);
-        return array($class, $method, null);
+        return [$class, $method, null];
     }
-    return array(null, $callback, null);
+    return [null, $callback, null];
 }
 
-function callbackTargetDefined($callback, $shouldAutoload = false)
+function callableDefined($callable, $shouldAutoload = false)
 {
-    list($class, $method, $instance) = interpretCallback($callback);
+    list($class, $method, $instance) = interpretCallable($callable);
     if ($instance !== null) {
         return true;
     }
@@ -94,12 +96,8 @@ function callbackTargetDefined($callback, $shouldAutoload = false)
 
 function classOrTraitExists($classOrTrait, $shouldAutoload = true)
 {
-    if (traitsSupported()) {
-        if (trait_exists($classOrTrait, $shouldAutoload)) {
-            return true;
-        }
-    }
-    return class_exists($classOrTrait, $shouldAutoload);
+    return class_exists($classOrTrait, $shouldAutoload)
+        || trait_exists($classOrTrait, $shouldAutoload);
 }
 
 function append(&$array, $value)
@@ -119,18 +117,73 @@ function reflectCallback($callback)
     if ($callback instanceof \Closure) {
         return new \ReflectionFunction($callback);
     }
-    list($class, $method) = interpretCallback($callback);
+    list($class, $method) = interpretCallable($callback);
     if (isset($class)) {
         return new \ReflectionMethod($class, $method);
     }
     return new \ReflectionFunction($method);
 }
 
-function callbackToString($callback)
+function callableToString($callback)
 {
-    list($class, $method) = interpretCallback($callback);
+    list($class, $method) = interpretCallable($callback);
     if (isset($class)) {
         return $class . "::" . $method;
     }
     return $method;
+}
+
+function alias($namespace, array $mapping)
+{
+    foreach ($mapping as $original => $aliases) {
+        $original = ltrim(str_replace('\\', '\\\\', $namespace) . '\\\\' . $original, '\\');
+        foreach ((array) $aliases as $alias) {
+            eval(sprintf(ALIASING_CODE, $namespace, $alias, $original));
+        }
+    }
+}
+
+function getUserDefinedCallables()
+{
+    # TODO optimize
+    return array_merge(get_defined_functions()['user'], getUserDefinedMethods());
+}
+
+function getUserDefinedMethods()
+{
+    # TODO optimize
+    $classes = array_merge(getUserDefinedClasses(), get_declared_traits());
+    $mapping = function($class) {
+        $methods = get_class_methods($class);
+        return array_map(function ($method) use ($class) {
+            return "$class::$method";
+        }, $methods);
+    };
+    return call_user_func_array('array_merge', array_map($mapping, $classes));
+}
+
+function getUserDefinedClasses()
+{
+    # TODO optimize
+    $all = get_declared_classes();
+    return array_filter($all, function($class) {
+        return (new \ReflectionClass($class))->isUserDefined();
+    });
+}
+
+function matchWildcard($wildcard, $subject)
+{
+    # TODO optimize
+    $table = ['*' => '.*', '{' => '(', '}' => ')', ' ' => ''];
+    return preg_match('/' . strtr($wildcard, $table) . '/', $subject);
+}
+
+function isOwnName($name)
+{
+    return stripos((string) $name, 'Patchwork\\') === 0;
+}
+
+function isForeignName($name)
+{
+    return !isOwnName($name);
 }
